Refactor folder/file structure based on atomin design system, added asset folder for common images and added bootstrap.
