+++ frontend/package.json
@@ -6,6 +6,7 @@
     "@storybook/preset-scss": "^1.0.3",
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
+    "bootstrap": "^5.2.3",
     "react": "^17.0.0",
     "react-dom": "^17.0.0",
     "recoil": "^0.7.7",
