+++ frontend/src/components/pages/Home/Page.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Header } from './Header';
+import { Header } from '../../organisms/Header/Header';
 import './page.css';
 
 export const Page = () => {
