+++ frontend/src/stories/Header.stories.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Header } from './Header';
+import { Header } from '../components/organisms/Header/Header';
 
 export default {
   title: 'Example/Header',
